Refactor Shelley Insert.hs
