+++ cardano-db-sync/cardano-db-sync.cabal
@@ -86,8 +86,14 @@ library
                         Cardano.DbSync.Era.Shelley.Generic.Witness
                         Cardano.DbSync.Era.Shelley.Genesis
                         Cardano.DbSync.Era.Shelley.Insert
+                        Cardano.DbSync.Era.Shelley.Insert.Certificate
                         Cardano.DbSync.Era.Shelley.Insert.Epoch
+                        Cardano.DbSync.Era.Shelley.Insert.GovAction
                         Cardano.DbSync.Era.Shelley.Insert.Grouped
+                        Cardano.DbSync.Era.Shelley.Insert.Other
+                        Cardano.DbSync.Era.Shelley.Insert.Pool
+                        Cardano.DbSync.Era.Shelley.Insert.Tx
+
                         Cardano.DbSync.Era.Shelley.Query
                         Cardano.DbSync.Era.Shelley.Validate
