+++ cardano-db-sync/src/Cardano/DbSync/Era/Shelley/Genesis.hs
@@ -17,7 +17,9 @@ import Cardano.DbSync.Api
 import Cardano.DbSync.Api.Types (SyncEnv (envBackend))
 import Cardano.DbSync.Cache.Types (Cache (..), uninitiatedCache)
 import qualified Cardano.DbSync.Era.Shelley.Generic.Util as Generic
-import Cardano.DbSync.Era.Shelley.Insert
+import Cardano.DbSync.Era.Shelley.Insert.Certificate (insertDelegation, insertStakeRegistration)
+import Cardano.DbSync.Era.Shelley.Insert.Other (insertStakeAddressRefIfMissing)
+import Cardano.DbSync.Era.Shelley.Insert.Pool (insertPoolRegister)
 import Cardano.DbSync.Era.Util (liftLookupFail)
 import Cardano.DbSync.Error
 import Cardano.DbSync.Util
